Check byte framing plus
When Arbotix Pro does not have power going out to servos and there are
servos connected get random garbage on buss.  Which at times crashed the
analyzer.  So I don't process a byte if we the start bit or stop bit are
not have the correct value...

Also attempt number 3 of getting export working (fail)