WIP - strcat_s -> std::stringstream, ...
Again WIP hopefully shows Results from Reads, as Words if appropriate.
It caches the Register Start from the previous Read request into a word
of the Results frame and again if appropriate displays words...

Also and maybe more important.  I removed the usage of snprintf plus
strcat, like calls and now instead use std::stringstream to build the
outputs.
This is what Mark of Saleae said they use in many of their analyzers to
get around some of the cross build issues as well as to reduce
probability of memory corruption.